--- src/getElements.js
@@ -1,1 +1,0 @@
-export const data = document.getElementById('data');
