+++ src/index.html
@@ -7,6 +7,22 @@
     <title>To Do list</title>
   </head>
   <body>
-    <ul id="data"></ul>
+    <header class="header">
+      <div class="to-do">
+        <h2 class="heading">Today's To Do</h2>
+      </div>
+    </header>
+    <section class="input-add">
+      <div class="input"></div>
+        <input type="text" placeholder="Add to your list..." />
+      </div>
+      </section>
+
+      <section class="do-list">
+        <ul id="data"></ul>
+        <div class="clear">
+          <button>Clear all completed</button>
+        </div>
+    </section>
   </body>
 </html>
