--- src/list.js
@@ -1,17 +1,0 @@
-const list = [
-  {
-    index: 0,
-    description: 'Send an email',
-    completed: false,
-  },
-  {
-    index: 1,
-    description: 'Go to Gym',
-    completed: true,
-  },
-  {
-    index: 2,
-    description: 'Call Mirza',
-    completed: false,
-  },
-];
