--- src/showList.js
@@ -1,3 +1,0 @@
-const ShowList = () => {};
-
-export default ShowList;
