+++ src/style.css
@@ -1,3 +1,20 @@
 body {
-  background-color: red;
+  background-color: rgb(255, 255, 255);
+  box-sizing: border-box;
+  margin: 10%;
+  padding: 0;
+  border: 1px solid #7e7e7e;
+}
+
+.heading {
+  font-size: 18px;
+  font-weight: 400;
+  margin: 10px;
+}
+ul {
+  border: 1px solid rgb(238, 238, 238);
+}
+.list-el {
+  display: flex;
+  gap: 10px;
 }
